python/requirements.txt
paho-mqtt
numpy
pandas
scikit-learn
matplotlib
